% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_budget.R
\name{annual_budget}
\alias{annual_budget}
\title{Downloads state budgets}
\usage{
annual_budget(year, abb)
}
\arguments{
\item{year}{year of the budget
abb  abbreviation of the state in either lower ou upper case.}
}
\description{
copyryig (c) 2016, José de Jesus Filho
last modified Nov, 2016
}
\details{
This program is free software; you can redistribute it and/or

This function downloads state budgets from 2000 to 2012.
}
\examples{
budget_sp_2000<-annual_budget(2000,"sp")
}
\keyword{budget}
\keyword{states}

