% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citiesBR.R
\name{citiesBR}
\alias{citiesBR}
\title{Get Brazilian state cities from IBGE}
\usage{
citiesBR(abb)
}
\arguments{
\item{abb}{abbreviation of the state}
}
\description{
copyryig (c) 2016, José de Jesus Filho
last modified Nov, 2016
}
\details{
This program is free software; you can redistribute it and/or

This function returns all cities of a Brazilian state according
     to the Brazilian Institute of Geography. You can write in
     either upper or lower case.
}
\examples{
sergipe<-citiesBR("se")
amapa<-citiesBR("AP")
}
\keyword{cities}
\keyword{municipalities}
\keyword{states}

