% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estados.R
\name{estados}
\alias{estados}
\title{Return a dataframe with Brazilian states' names, their abbreviation and their capitals}
\usage{
estados(x)
}
\arguments{
\item{...}{}
}
\description{
copyryig (c) 2016, José de Jesus Filho
last modified Nov, 2016
 
 This program is free software; you can redistribute it and/or
}
\details{
This is a simple function that return a dataframe with the names of the Brazilian states,
    their abbreviations and their capitals.
}
\examples{
df<-estados()

}
\keyword{Brazilian}
\keyword{abbreviations}
\keyword{states}

