% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_br.R
\name{number_br}
\alias{number_br}
\title{Convert messy Brazilian "numeric" strings to numbers}
\usage{
number_br(str)
}
\arguments{
\item{str}{strings vector with messy "numeric" strings}
}
\description{
copyryig (c) 2016, José de Jesus Filho
    last modified Nov, 2016
}
\details{
This program is free software; you can redistribute it and/or

This function converts messy "numeric" strings such as "346.155,7   " to numbers:
    346155.7
}
\examples{
number_br("346.155,7      ")

}
\keyword{numbers}
\keyword{parse}
\keyword{strings}

