% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postal.R
\name{postal}
\alias{postal}
\title{Find addresses based on the Brazilian postal code numbers}
\usage{
postal(cep)
}
\arguments{
\item{cep}{the postal code}
}
\description{
copyryig (c) 2016, José de Jesus Filho
last modified Nov, 2016
}
\details{
This program is free software; you can redistribute it and/or

This function returns a data.frame with six columns: cep (postal code),
     tipoDeLogradouro (type of address), logradouro (address),
     bairro (neiborhood), cidade (city), estado (state).
     If you add non numeric caracteres, the function will parse it
     to number and pad with zeros on the left side.
}
\examples{
address1<-postal("02828-030")
address2<-postal(02828030)
}
\keyword{address}
\keyword{cep}
\keyword{code}
\keyword{postal}

